requirements.txt
matplotlib==2.2.2
numpy==1.14.3
pandas==0.22.0
Pillow==5.1.0
scikit-image==0.13.1
scikit-learn==0.19.1
scipy==1.1.0
tensorboard==1.8.0
tensorflow==1.8.0
tensorflow-gpu==1.8.0
torch==0.4.1
torchvision==0.2.1

